modified:   Doc/readme.txt 	modified:   User/_apollorobot.c 	modified:   User/_apollorobot.h 	modified:   User/app/app.c 	modified:   User/bsp/adc/bsp_adc.c 	modified:   User/bsp/adc/bsp_adc.h 	modified:   User/bsp/bsp.c 	modified:   User/bsp/bsp.h 	modified:   User/bsp/digitalsensor/bsp_digitalsensor.c 	modified:   User/bsp/ultrasnio/bsp_ultrasnio.c 	modified:   User/bsp/ultrasnio/bsp_ultrasnio.h 	modified:   User/font/SongTi12.c 	modified:   User/window/WIDGET_Instructor.c 	modified:   User/window/Window_1_1.c 	modified:   User/window/Window_Running.c

+++ User/bsp/adc/bsp_adc.h
@@ -12,7 +12,7 @@
 
 
 extern void adc_Init(void);
-extern u16 Get_adc(u8 ch);
+extern float Get_adc(u8 ch) ;
 extern u16 Get_Adc_Average(u8 ch,u8 times);
 extern int Get_Temp(void);
